requirements.txt
discord.py-self
aiohttp
requests
pycryptodome
pywin32